+++ sources/packages/backend/apps/api/src/testHelpers/index.ts
@@ -6,3 +6,4 @@ export * from "./mocked-providers/jwt-service-mock";
 export * from "./testing-modules/testing-modules-helper";
 export * from "./auth";
 export * from "./fake-entities";
+export * from "./program-year/program-year-helper";
